+++ R/analyze.lm.R
@@ -58,7 +58,7 @@ analyze.lm <- function(x, CI=95, ...) {
 
   if (!is.null(CI)) {
     CI_values <- confint(fit, level = CI / 100)
-    CI_values <- tail(CI_values, n = 2)
+    CI_values <- tail(CI_values, n = length(rownames(fitsum)))
     fitsum$CI_lower <- CI_values[, 1]
     fitsum$CI_higher <- CI_values[, 2]
   }